requirements.txt
pandas
numpy
PyYAML
scikit-learn
joblib
matplotlib
plotly
streamlit
ta
MetaTrader5
lightgbm
